requirements.txt
Appium-Python-Client==1.1.0
DateTime==4.3
logger==1.4
Pillow==9.2.0
pytz==2021.1
selenium==3.141.0
thread6==0.2.0
urllib3==1.26.4
zope.interface==5.4.0
playsound==1.3.0
xlrd==1.2.0
XlsxWriter==1.3.7
psutil==5.8.0
